% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlMoonPlot.R
\name{moonplot}
\alias{moonplot}
\title{rhtmlMoonPlot HTML Widget}
\source{
https://github.com/Displayr/rhtmlMoonPlot
}
\usage{
moonplot(settingsJsonString = "{}")
}
\arguments{
\item{param1}{is a good param}
}
\description{
A HTMLWidget that renders moonplots
}
\section{Usage Scenarios}{


Scenario 1: Blah blah
}

\examples{

rhtmlMoonPlot::moonplot('{}')

}
\author{
Displayr <opensource@displayr.com>
}
