% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlMoonPlot.R
\name{moonplot}
\alias{moonplot}
\title{rhtmlMoonPlot}
\source{
https://github.com/NumbersInternational/rhtmlMoonPlot
}
\usage{
moonplot(coreNodes = NULL, surfaceNodes = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{coreNodes}{Coordinates of nodes in the center of the moon}

\item{surfaceNodes}{Coordinates of nodes outside of the moon}

\item{width}{Width of the viewer window}

\item{height}{Height of viewer window}
}
\description{
Uses correspondence analysis to create a Moon Plot
}
\author{
Po Liu <pliu0771@uni.sydney.edu>
}

