% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlMoonPlot.R
\name{moonPlot}
\alias{moonPlot}
\title{rhtmlTemplate HTML Widget}
\source{
https://github.com/Displayr/rhtmlTemplate
}
\usage{
moonPlot(settingsJsonString = "{}")
}
\arguments{
\item{param1}{is a good param}
}
\description{
A HTMLWidget that ...
}
\section{Usage Scenarios}{


Scenario 1: Blah blah
}

\examples{

rhtmlTemplate::template('{}')

}
\author{
First Last <first.last@gmail.com>
}
